+++ core/src/components/datetime/datetime-interface.ts
@@ -5,3 +5,13 @@ export interface DatetimeOptions {
 export interface DatetimeChangeEventDetail {
   value?: string | null;
 }
+
+export interface DatetimeParts {
+  month: number;
+  day: number | null;
+  year: number;
+  dayOfWeek?: number | null;
+  hour?: number;
+  minute?: number;
+  ampm?: 'am' | 'pm';
+}
