+++ core/src/components/datetime/datetime.ios.scss
@@ -3,11 +3,8 @@
 @import "../../themes/ionic.globals.ios";
 
 :host {
-  --background: var(--ion-color-step-100, #ffffff);
-
-  display: flex;
-
-  flex-flow: column;
+  --background: var(--ion-color-light, #ffffff);
+  --background-rgb: var(--ion-color-light-rgb);
 
   width: 316px;
 }
@@ -24,25 +21,6 @@
   color: $text-color-step-400;
 
   font-size: 14px;
-}
-
-// Calendar
-// -----------------------------------
-
-/**
- * This allows the calendar to take
- * up 100% of the remaining height.
- * On iOS, if there are more than
- * 5 rows of dates, the dates should
- * be resized to fit into this
- * container.
- */
-:host .datetime-calendar {
-  display: flex;
-
-  flex: 1 1 auto;
-
-  flex-flow: column;
 }
 
 // Calendar / Header / Action Buttons
@@ -240,6 +218,132 @@
 
 :host .time-item {
   line-height: $datetime-ios-time-height;
+}
+
+// Year Picker
+// -----------------------------------
+:host(.show-month-and-year) .calendar-action-buttons ion-item {
+  --color: #{current-color(base)};
+}
+
+:host .datetime-year-body .datetime-picker-col {
+  @include margin(0, 10px, 0, 10px);
+  @include padding(0, $datetime-ios-padding, 0, $datetime-ios-padding);
+}
+
+:host .datetime-picker-before {
+  @include position(0, null, null, 0);
+
+  position: absolute;
+
+  width: 100%;
+
+  height: 82px;
+
+  background: linear-gradient(to bottom, var(--background, var(--ion-background-color, #fff)) 20%, rgba(var(--background-rgb, var(--ion-background-color-rgb, 255, 255, 255)), 0.8) 100%);
+
+  z-index: 10;
+
+  pointer-events: none;
+}
+
+:host .datetime-picker-after {
+  @include position(116px, null, null, 0);
+
+  position: absolute;
+
+  width: 100%;
+
+  height: 115px;
+
+  background: linear-gradient(to top, var(--background, var(--ion-background-color, #fff)) 30%, rgba(var(--background-rgb, var(--ion-background-color-rgb, 255, 255, 255)), 0.8) 100%);
+
+  z-index: 10;
+
+  pointer-events: none;
+}
+
+:host .datetime-picker-highlight {
+  @include position(50%, 0, 0, 0);
+  @include border-radius($datetime-ios-time-border-radius, $datetime-ios-time-border-radius, $datetime-ios-time-border-radius, $datetime-ios-time-border-radius);
+  @include margin(0, auto, 0, auto);
+
+  position: absolute;
+
+  width: calc(100% - #{$datetime-ios-padding * 2});
+
+  height: 34px;
+
+  transform: translateY(-50%);
+
+  background: var(--ion-color-step-150, #eeeeef);
+
+  z-index: -1;
+}
+
+:host .datetime-year-body {
+  display: flex;
+
+  position: relative;
+
+  align-items: center;
+
+  justify-content: center;
+
+  font-size: 22px;
+}
+
+:host .datetime-picker-col {
+  scroll-snap-type: y mandatory;
+
+  /**
+   * Need to explicitly set overflow-x: hidden
+   * for older implementations of scroll snapping.
+   */
+  overflow-x: hidden;
+  overflow-y: scroll;
+
+  // Hide scrollbars on Firefox
+  scrollbar-width: none;
+
+  height: 200px;
+
+  outline: none;
+}
+
+@media (any-hover: hover) {
+  :host .datetime-picker-col:focus {
+    background: current-color(base, 0.2);
+  }
+}
+
+/**
+ * Hide scrollbars on Chrome and Safari
+ */
+:host .datetime-picker-col::-webkit-scrollbar {
+  display: none;
+}
+
+
+:host .picker-col-item {
+  height: 34px;
+
+  line-height: 34px;
+
+  scroll-snap-align: center;
+}
+
+:host .picker-col-item-empty {
+  scroll-snap-align: none;
+}
+
+
+:host .datetime-year-body .datetime-picker-col:first-of-type {
+  text-align: left;
+}
+
+:host .datetime-year-body .datetime-picker-col:last-of-type {
+  text-align: right;
 }
 
 // Footer
