+++ core/src/components/datetime/datetime.md.scss
@@ -19,15 +19,15 @@
 }
 
 :host .datetime-header .datetime-title {
-  font-size: 12px;
+  font-size: $datetime-md-title-font-size;
 
   text-transform: uppercase;
 }
 
 :host .datetime-header .datetime-selected-date {
-  margin-top: 30px;
+  @include margin(30px, null, null, null);
 
-  font-size: 35px;
+  font-size: $datetime-md-selected-date-font-size;
 }
 
 // Calendar / Header / Action Buttons
@@ -48,7 +48,7 @@
 
   color: $text-color-step-500;
 
-  font-size: 13px;
+  font-size: $datetime-md-calendar-item-font-size;
 
   line-height: 36px;
 }
@@ -69,10 +69,11 @@
 
 }
 
+// Individual day button in month
 :host .calendar-day {
   @include padding(13px, 0, 13px, 0px);
 
-  font-size: 14px;
+  font-size: $datetime-md-calendar-item-font-size;
 }
 
 :host .calendar-day:focus:after {
@@ -151,6 +152,97 @@
   background: var(--background-checked);
 }
 
+// Year Picker
+// -----------------------------------
+:host(.show-month-and-year) .datetime-calendar {
+  flex: 0;
+}
+:host(.show-month-and-year) .datetime-year {
+  flex: 1;
+
+  min-height: 0;
+}
+
+:host(.show-month-and-year) .datetime-footer {
+  border-top: 1px solid var(--ion-color-step-250, #dddddd);
+}
+
+:host .datetime-year-body {
+  @include padding(0, $datetime-md-padding, $datetime-md-padding, $datetime-md-padding);
+
+  display: grid;
+
+  grid-template-columns: repeat(auto-fit, minmax(74px, 1fr));
+  grid-gap: 0px 6px;
+
+  overflow-y: scroll;
+  -webkit-overflow-scrolling: touch;
+}
+
+:host .datetime-year-item {
+  @include padding(0px, 0px, 0px, 0px);
+  @include margin(0px, 0px, 0px, 0px);
+
+  position: relative;
+
+  align-items: center;
+  justify-content: center;
+
+  border: none;
+
+  outline: none;
+
+  background: none;
+  color: currentColor;
+
+  cursor: pointer;
+
+  appearance: none;
+
+  z-index: 0;
+}
+
+:host .datetime-year-item[disabled] {
+  pointer-events: none;
+
+  opacity: 0.4;
+}
+
+:host .datetime-year-item .datetime-year-inner {
+  @include border-radius(20px, 20px, 20px, 20px);
+  @include margin(10px, 10px, 10px, 10px);
+
+  display: flex;
+
+  align-items: center;
+  justify-content: center;
+
+  height: 32px;
+
+  border: 1px solid transparent;
+
+  font-size: 16px;
+}
+
+:host .datetime-current-year .datetime-year-inner {
+  border: 1px solid current-color(base);
+
+  color: current-color(base);
+}
+
+:host .datetime-active-year .datetime-year-inner {
+  border: 1px solid current-color(base);
+
+  background: current-color(base);
+  color: current-color(contrast);
+}
+
+@media (any-hover: hover) {
+  :host .datetime-year-item:hover,
+  :host .datetime-year-item:focus {
+    background: current-color(base, 0.1);
+  }
+}
 
 // Footer
 // -----------------------------------
