+++ core/src/components/datetime/datetime.md.vars.scss
@@ -1,17 +1,26 @@
 // MD Datetime
 // --------------------------------------------------
 
+/// @prop - Font size for title in header
+$datetime-md-title-font-size:          12px !default;
+
+/// @prop - Font size for selected date in header
+$datetime-md-selected-date-font-size:  34px !default;
+
+/// @prop - Font size for calendar day button
+$datetime-md-calendar-item-font-size:  14px !default;
+
 /// @prop - Padding for content in header
-$datetime-md-header-padding:       20px !default;
+$datetime-md-header-padding:           20px !default;
 
 /// @prop - Padding for content
-$datetime-md-padding:              16px !default;
+$datetime-md-padding:                  16px !default;
 
 /// @prop - Height of the time picker
-$datetime-md-time-height:          28px !default;
+$datetime-md-time-height:              28px !default;
 
 /// @prop - Width of the time picker
-$datetime-md-time-width:           68px !default;
+$datetime-md-time-width:               68px !default;
 
 /// @prop - Border radius of the time picker
-$datetime-md-time-border-radius:   4px !default;
+$datetime-md-time-border-radius:       4px !default;
