+++ core/src/components/datetime/test/demo/index.html
@@ -243,7 +243,7 @@
           </ion-item>
         </div>
 
-        <ion-datetime>
+        <ion-datetime max="2023-10-15" min="2015-07-04">
           <span slot="title">Select Date</span>
           <ion-buttons slot="buttons">
             <ion-button color="primary">Cancel</ion-button>
